@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DataFormComponent } from './data-form/data-form.component';
 import { FormulariosRoutingModule } from './formularios-routing.module';
@@ -9,19 +9,13 @@ import { FormulariosComponent } from './formularios.component';
 import { SharedModule } from './shared/shared.module';
 import { TemplateFormComponent } from './template-form/template-form.component';
 
-@NgModule({
-  declarations: [
-    FormulariosComponent,
-    TemplateFormComponent,
-    DataFormComponent,
-  ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    FormulariosRoutingModule,
-    HttpClientModule,
-    SharedModule,
-  ],
-})
+@NgModule({ declarations: [
+        FormulariosComponent,
+        TemplateFormComponent,
+        DataFormComponent,
+    ], imports: [CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        FormulariosRoutingModule,
+        SharedModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class FormulariosModule {}
